Move find_versions to Source, and differentiate between ARPA and ARPA Git. refs: #101
